refactor: Rule에 어댑터 패턴 적용
